requirement.txt
python-decouple
djangorestframework
django-redis